[expirer] skip too big task_containers
There are soo big, that they slow down the expiry of others by several month now.
So the strategy is to skip them in expirer, clean them up manually and
afterwards delete the target objects manually.

This 2 explictly listed task_containers (timestamps) are not existing in other spot
checked regions.

//cc @majewsky